#50 refactor: _components から _stores への import を集約 index 経由に統一
- 個別 store path (../../_stores/actor 等) を ../../_stores 経由に統一
- 対象: action-bar, action-log-panel, actor-controller, current-position-marker, planned-path-marker, schedule-preview
