+++ src/prototypes/time-control/time-control-03/_components/current-position-marker/index.tsx
@@ -3,8 +3,7 @@ import { memo, useEffect, useRef } from 'react'
 import { getTickMs } from '../../../time-control-02/_lib/get-tick-ms'
 import { useStageTransform } from '../../_contexts/stage-transform-context'
 import { toPixelStyle } from '../../_lib/stage-coords'
-import { useActorStore } from '../../_stores/actor'
-import { usePositionStoreApi } from '../../_stores/position'
+import { useActorStore, usePositionStoreApi } from '../../_stores'
 import { ActorId, Position } from '../../types'
 
 type CurrentPositionMarkerProps = {
